RD-1419 Save exec-group results into dep-group (#3005)
* relationships: allow unique for m2m

unfortunate that we didnt make the m2m tables unique. It means that
eg. a deployment can belong to a group multiple times, a user
can belong to a group multiple times, etc.
The ORM kinda protects against that, but it'd be better if the DB
itself did.

* Make the dep-groups/deps m2m unique

* Add .success/failed target dep groups for exec groups

* Migrations

Those are for the previous 2 commits: the unique stuff, and the
target dep group FKs

* PUT /exec-groups/id to set success/failure groups

When setting a group, all already-finished deps will be put
into it right then and there.

* When finishing an execution, put the dep into the target group

* tests

several cases per test, because i only want to create the deployment
once, and still be able to test multiple things

* replace put with patch