tests_executions: add more wait-for-execution
Similar to #3018 - waiting for executions to end.

With this, every test in test_executions.py, cleans up after itself,
ie. waits for all of its exeuctions to end.

Otherwise, those executions will be left running in the mgmtworker,
but cleaned up from the db, which will just result in 401s in the log,
when the mgmtworker tries to actually run them. Ugly.

(and no, we can't "clean" the mgmtworker, there's no way to do that)