UTCDateTime: pass a tz-naive datetime back to the db
We get a datetime from the db, we add the string 'Z' to it, making a
string representation that does contain a timezone; so going back,
we must mirror that and strip it back.

Also, note that this can be either a string (in case it was retrieved
from the db) or a datetime object (in case it was just assigned),
so handle both. This is because we for some reason love strings
rather than actual types. But we'll get to eliminating silly stuff
like .get_formatted_timestamp in due time. Not yet. For now, handle both.