RD-2273 queued exec inte-tests: handle changed timezones, 6.0.0 (#3028)
* .all() for the get-queued query

explicitly run the query right there, rather than implicitly via
iterating over it. Mostly doesn't matter, helps with debugging.
Might well be explicit

* UTCDateTime: pass a tz-naive datetime back to the db

We get a datetime from the db, we add the string 'Z' to it, making a
string representation that does contain a timezone; so going back,
we must mirror that and strip it back.

Also, note that this can be either a string (in case it was retrieved
from the db) or a datetime object (in case it was just assigned),
so handle both. This is because we for some reason love strings
rather than actual types. But we'll get to eliminating silly stuff
like .get_formatted_timestamp in due time. Not yet. For now, handle both.

* EventsAlternativeTimezoneTest: set the db tz back

these tests change the database's tz, so it would be nice if they
cleaned up after themselves and set it back.

* flake