snapshot-restore: also pause the execution-scheduler, 6.0.0 (#3031)
* snapshot-restore: also pause the execution-scheduler

Similar to pausing amqp-postgres, those db-using services must not
run while the db is being down/upgraded. The scheduler even more so,
because if it just happens to try and run something during that time,
oh boy is that going to break.

* inte-tests events printer: retry on any amqp exception

we restart rabbitmq and do all kinds of things in the tests, the
amqp connection might break in many ways. I just give up and retry on
anything.