inte-tests: wait for executions to finished in the 2-scheduled test (#3036)
* inte-tests: wait for executions to finished in the 2-scheduled test

this tests schedules 2 executions, waits for them to start, and
finishes. Instead, it must wait for them to end. If it doesn't, then
they might just start installing the plugin, the next test will see
"oh the plugin is being installed, ok", this already-finished test
will stop due to the db cleanup, all kinds of weird interactions.
Let's just avoid all of that.

* same in these tests