RD-523 Optimize the operation-update call (#3053)
* Operation&TasksGraph models: default id to uuid

Those items always use uuids as their id; and so, they also can
have is_id_unique=False (yes I know it seems backwards, but that's
what it does), to avoid the expensive conflict check. No need to check
conflicts on uuids.

* Put creating the tasksgraph under a single transaction

No reason to commit every operation separately. This is way faster.

* Update operation state: use a SQL UPDATE for increasing op counts

For increasing the count on the execution object, use a single atomic
UPDATE call rather than a fetch-then-update. This is a bit faster.

* Also do the same for the delete path

Deleting operation isn't really a thing that needs to be fast,
but might as well use the same code, why not.
Of course this also changes the name of the method because it's not
"increase" anymore, is it.