inte-tests config: add some retries
In inte-tests, we set the manager config to have 0 retries and 0
retry-interval; this saves some time on failing tests (don't want
to wait the default 15 minutes), but is too unrealistic. Actual
usage of the manager will of course always have some retries.

These settings try to be more realistic, while still being as fast
as possible.