inte-tests: a few more cleanups (#3080)
* workflows tests: synchronize the operations

those workflows used to just send operations and never wait for them,
and that leads to a ton of races. Instead, wait for every operation
as we send it.

* remove unused method

* remove unused host-related agent test methods

those are a leftover of the times when we had several docker hosts
with docl and dockercompute and whatever else.
Anyway, now, agents are installed just on the manager container,
and so all of this is unnecessary.

* remove _get_latest_execution

another unused method

* remove unused _wait_for_rest

yet another unused method: we have wait-for-starter instead now

* remove patch_yaml

tbh i dont know what was this even used for. Something docl probably.