RD-2486 Throw an error if the blueprint dir already exists
This ports #3107 to 5.2.3

* use os.rename rather than shutil.move for the last move

this last rename of the dir is always inside of a single filesystem,
under /opt/manager/resources, so we can use os.rename directly

the difference is if the target directory already exists:
* target/ doesnt exist:
    * os.rename just renames the dir
    * so does shutil.move, and indeed it does use os.rename under
      the hood
* target/ exists and is empty
    * os.rename just stomps over the empty dir
    * so does shutil.move
* target/ exists and is not empty
    * os.rename throws an OSError
    * shutil.move moves the source dir INTO the target dir, so that
      we have target/source/   - THAT IS WRONG!

That last part is not what we want, because then we end up with
blueprints in an additional subdirectory, and then, of course scripts
and files from the blueprint can't be found under their expected
locations.

This happens in a cluster, with replication lag, where you had
a blueprint that you removed, and you upload a blueprint with the
same name. Then the old directory might have not been removed yet
on all the cluster nodes.

* Translate that error to a blueprint error

When the error from the first commit is thrown, make sure to mark
the blueprint as failed.

* tests: clean the tmpdir after each test

similar to how we drop the db, also the tmpdir needs to be cleaned.
Weird that it never caused any problems before!