Skip role checking in a simple case
If we're already authorized as an execution, allow some operations
implicitly: sending events, and updating operations. The idea
is that those are going to be allowed for any execution to do,
and that if you were allowed to start an execution, then you must
necessarily be also allowed to do those two things.

This saves a LOT of time, because checking roles is pretty slow.