RD-2659 Optimize getting current_execution
storage-manager is just unnecessary here, although that's not the
most important change.

Join tenant and creator, so that when the code next does
current_execution.tenant, and current_execution.creator (for the
purposes of set_current_tenant/set_current_user), those items
are already there, and don't require another query.