snap-res: restore current execution by appending the dump
We must avoid having the db in a state in which we can't send in
requests. If we restore the dump first, and only then run the
restore-current-execution query, then in the time between those
two calls, requests can't be run. And since the events/logs are
a rest request now, any `ctx.logger` call between those two, will fail.

To avoid that, just restore the current execution in the same call
as restoring the dump.