remove many debug logs
this is unfortunate and i wish i didnt have to do this :(

Anyway, now that logs are sent in synchronously, we can't send them
in while the db is downgraded, because in that time, the restservice
won't be able to process those requests.

So we'll just lose those logs. Not the best thing to do, but
I'm gonna say it's better than the alternative, which is have it
not work at all.

And I mean, those are just debug logs, and all of this is HOPEFULLY
on its way out soon!