minimize the time that the db is downgraded
it only needs to be downgraded for this very call, not for the whole
method. If I didn't do this, then I'd have to remove even more logs,
and even more concurrently-running requests (eg. monitoring) would
fail.

This mostly reverts #3144 - but the logic is similar, we want to spend
as little time as possible in a state where the db is unusable.
However, the execution must be restored AFTER we go back to the current
db schema version - because if we downgrade to eg. 4.4, then we
can't restore the execution by appending to the dump, because
the schema is different - eg. the token column doesnt exist