snapshot-restore: avoid using the db when downgraded (#3149)
* config last-updated query: change the order of the parts

the query doesn't care which one is first - config or roles, but
we always unconditionally lock the roles table in snap-res.
And if config is first, that's not locked by snap-res if the
snapshot is older than 4.6, and so there might be a deadlock between
this, and the snapshot itself

* follow-events.py: limit the size of the event

postgres has a hard limit, which seems to happen to be 8000 (in some
versions?)
anyway, 6000 works.

The events that don't fit, just won't be shown in the inte-tests
log follow. Oh well, too bad for them.

* remove many debug logs

this is unfortunate and i wish i didnt have to do this :(

Anyway, now that logs are sent in synchronously, we can't send them
in while the db is downgraded, because in that time, the restservice
won't be able to process those requests.

So we'll just lose those logs. Not the best thing to do, but
I'm gonna say it's better than the alternative, which is have it
not work at all.

And I mean, those are just debug logs, and all of this is HOPEFULLY
on its way out soon!

* minimize the time that the db is downgraded

it only needs to be downgraded for this very call, not for the whole
method. If I didn't do this, then I'd have to remove even more logs,
and even more concurrently-running requests (eg. monitoring) would
fail.

This mostly reverts #3144 - but the logic is similar, we want to spend
as little time as possible in a state where the db is unusable.
However, the execution must be restored AFTER we go back to the current
db schema version - because if we downgrade to eg. 4.4, then we
can't restore the execution by appending to the dump, because
the schema is different - eg. the token column doesnt exist

* restore the license ASAP

similar to the previous commit - minimize the time in which the
system (in this case no longer the db - but the restservice)
is unusable. After this is done, restservice works again, and
ctx.logger will work again.

* Add some logs

I've removed a metric ton of ctx.logger.debug calls, because they
can't be under the db-schema-downgrade block. So let's at least
send a log when that critical block begins and ends.

* remove the idea of "protected_query"

we don't log queries anymore. Because we can't really do that, as noted
in a previous commit. So this whole this is unnecessary.
And maybe it was a bad idea anyway. Who knows!

* flake