amqp-postgres: ignore hook messages
ever since cloudify-cosmo/cloudify-common#830 we don't send a hook
together with the actual event text message. The hook is sent
separately by the restservice, and the event text message is
created by the mgmtworker.
This led to the message being printed twice, once by the mgmtworker's
event, and once by the amqp-postgres catching the hook. So let's have
it not print the hook.