sm._get_base_query: rewrite the include stuff
before, passing include meant that we returned sqlalchemy keyedtuples
that contained just the requested attributes. Those keyedtuples
have the same interface, effectively - just attribute access
(ie. you can do `x.id` whether it's an actual model class, or just
a keyedtuple with only a few fields). That's why we were able to do
it.

The problem with that old approach however is that relationships
couldn't be loaded that way.

Instead, always load actual models, but only load the attributes
that were requested, using load_only.
And if relationships were requested as well, join them. We know they'll
have to be loaded, so we might as well join them up front, to avoid
running 1+n queries when rendering.

(and automatically include tenant_id, because it's always going
to be used internally)