RD-2773 Improve deployments-list performance (#3168)
* marshal_with: pass fields to to_response

model.to_response will now know what fields to include in the
response. This way, if `?_include=xxx` was passed in by the user,
to_response will no longer need to load all attributes, and some
can be skipped. Before, .to_response always loaded all attributes,
commonly running 1+n queries when loading relationships.

Note: when not passing include in the url params, this does mean that
all .response_fields are passed in. (not none, not empty list)

* model.to_response: only include the requested fields

the additional fields added in each model's .to_response, only add
them if they were actually asked for

This is what actually uses the previous commit :)

* models: cache .response_fields

computing those fields takes quite some time, because it examines
each attribute for its ORM metadata. Caching it makes rendering
responses way faster.

* sm._get_base_query: rewrite the include stuff

before, passing include meant that we returned sqlalchemy keyedtuples
that contained just the requested attributes. Those keyedtuples
have the same interface, effectively - just attribute access
(ie. you can do `x.id` whether it's an actual model class, or just
a keyedtuple with only a few fields). That's why we were able to do
it.

The problem with that old approach however is that relationships
couldn't be loaded that way.

Instead, always load actual models, but only load the attributes
that were requested, using load_only.
And if relationships were requested as well, join them. We know they'll
have to be loaded, so we might as well join them up front, to avoid
running 1+n queries when rendering.

(and automatically include tenant_id, because it's always going
to be used internally)

* keep include, even if labels are requested

now that include can handle relationships (since the previous commit),
we are able to keep using include, even if a relationship (labels)
is requested.

* simplify rest_utils functions

no need to mangle include anymore. And they can be simpler!

* sm.list: add load_relationships

add yet another flag! this will join all relationships declared
by the model.

We can't just `lazy='joined'` on every relationship, because for
many queries we don't actually want that (and it'd break locking)

* implement autoloaded relationships for deployments

* make include optional

after all, make it default to none. None means that all fields are
included. I didn't really want to do that, but there's a lot of
calling `.to_response()` with no arguments around the code :/
So this is for keeping compat with those places for now

* rewrite test_fields_query

now it returns something that has those attributes, ie. that actual
model, but it won't have those attributes loaded yet - assert that
they aren't loaded

* fixup

* test_blueprint_update_state: expunge the old blueprint

otherwise the update also updates the previous object

* agent.to_response: rabbitmq_X are not always there

* include now returns a model, doesnt have [0]

* just skip fields that dont have a .prop

those are things like labels or something. If we cant join them, well,
dont join them. They'll have to be retrieved on access.

* test filters: just check the deployment id

we want to see if the deployment is the same, just check its id.
This assert can fail otherwise because the labels list might be
ordered differently

* also fixup the interface in management_models

* default include to self.resource_fields/self.response_fields

* execgroups: reverse skip_fields

yeah, that's how it's supposed to be. Skip them if not get data.