test_dry_run: rewrite the asserts
those are the current messages. I've made them into regexps because
the subgraph name is in there, and that depends on the node-instance
id.
It's unfortunate that they're tied so closely to the actual text.
Or perhaps it's good. Anyway, that's what it used to be before,
and so it's still like that.