RD-2891 Operation events: also store events for subgraphs (#3175)
* operation events: add operation type to the format_event_message call

format_event_message now uses the type to know whether to say "Task"
or "Subgraph"

* operation events: also emit events for subgraph

...and no, subgraphs dont have context in them

* test_dry_run: rewrite the asserts

those are the current messages. I've made them into regexps because
the subgraph name is in there, and that depends on the node-instance
id.
It's unfortunate that they're tied so closely to the actual text.
Or perhaps it's good. Anyway, that's what it used to be before,
and so it's still like that.

* flake

* drop unnecessary escapes