mgmtworker workflow consumer: don't update operation state (#3177)
This deals with workflows/executions, not with operations. It ends up
trying to fetch the operation, using the execution id instead (haha),
gets a 404, and carries on, assuming the operation isn't stored.

I mean, it's not an operation. Just don't do it. Override the
method to do nothing instead.

A better solution would probably be to not use a inheritance hierarchy
here (why is a workflow consumer a child of an operation consumer?)
but for now, this removes the 1-per-execution 404 call in the rest logs