RD-2900 Allow using latest_exec_status in _include (#3179)
* Default latest_execution to the create execution when making a dep

Might as well. Status stays the same then. Both fields were added
in the same migration, so we don't even need to consider defaulting
one to the other in snapshots or anything, because if one is null,
then the other is null anyway

* Replace dep.latest_execution_status with a assoc_proxy

This will allow using it in _include etc other orm-related places.

Anyway, now the attribute itself is going to be returning the status
directly (eg. terminated/pending/started), and only for display,
we'll be transforming it into the DeploymentState thing (eg. completed/
in progress)