deployments v3: dont use super
in general, super() in those view functions is clumsy, because it
gives us back a serialized model already. Unfortunate. Let's skip
back to the regular sm.get() instead.
This is now made simpler, because dep v1 now just happens to be
a straight sm.get as well, without the hack that it used to have