RD-2916 deployments GET: automatically include `id` if needed (#3184)
* deployments v1: drop the now-useless hack

now that we include labels/etc properly, no need for this anymore

* deployments v3: dont use super

in general, super() in those view functions is clumsy, because it
gives us back a serialized model already. Unfortunate. Let's skip
back to the regular sm.get() instead.
This is now made simpler, because dep v1 now just happens to be
a straight sm.get as well, without the hack that it used to have

* serialize the deployment here

now that we dont use super, we get a Deployment object here, but
the ._populate method expects a dict, so let's make a dict. But now,
it will be a dict that obeys the _include

* automatically include id if needed