tests: create the database not in transaction
`CREATE DATABASE` cannot be run in a transaction, but now psycopg's
connect makes one implicitly when used as a ctxmanager. This is
indeed what they suggest instead!

Also, I've learned that the ctxmanager is _just_ for transaction
management, and it used to be for... nothing, before. It didn't
actually close the connection before. Maybe that's why we've seen
tests run out of postgres connections.