RD-2941 Upgrade sqlalchemy to 1.4 (#3196)
* upgrade sqlalchemy & psycopg

setup.py really isnt the right place, but for now, i'm not going to
make a proper reqs.txt just yet, in this pr

* rest_decorators: no need to specialcase sqlalchemy_collection

the private import doesn't exist anymore in 1.4.
However, since the include changes, we never return keyedtuples
anyway, so there's no need to have them!

* add an explicit .select() to the filter subqueries

sqlalchemy now wants a select object back, not a subquery object
(which is then implicitly converted to select)
This avoids many warnings

* declare the typedecorators as cache_ok

otherwise sqlalchemy throws warnings. And they are cache ok, because
they're just regular types, but with some added behaviour.

* node.instances: only have a backref

we used to have both a backref (implicitly in one_to_many_rel) and
an explicit relationship. That is now a warning in sqlalchemy. Keep
just the backref.

* add integrityerror to sql_errors

it is a sql error! and in new psycopg, commit-time check violation
is an integrityerror instead

* don't attempt to join with empty argument

that is actually now an error in sqlalchemy

* storage-manager: better form of logger.debug calls

not only does rendering the object actually format a string, it can
also run a query, if some attributes weren't loaded

* test_executions: create objects in a transaction

yeah this was more correct anyway. We touch multiple objects, we don't
necessarily want autocommit.

* tests: create the database not in transaction

`CREATE DATABASE` cannot be run in a transaction, but now psycopg's
connect makes one implicitly when used as a ctxmanager. This is
indeed what they suggest instead!

Also, I've learned that the ctxmanager is _just_ for transaction
management, and it used to be for... nothing, before. It didn't
actually close the connection before. Maybe that's why we've seen
tests run out of postgres connections.