RD-2832 Don't overwrite the RabbitMQ user on agent_create (#3204)
In case `agent_create` request comes in with a `create_rabbitmq_user`
flag enabled, we attempt to create a separate RabbitMQ user for that
agent.  In case the account already existed, we would have overwritten
it thus breaking the tenant's or the manager's interaction with the
broker.  This patch prevents that from happening.