RD-3006 Create a dep-update object in the workflow
Just adding to the skeleton.

I'm going to organize this workflow in a graph, so that all the
potentially-expensive computations can be stored and resumed if
cancelled, and in the future also replaced by remote tasks,
so that they can be split between nodes.

After this, there's going to be a second graph as well, for all the
installation.