REST: make the workflows part of the deployment response make sense
the workflow objects inside of deployments don't even contain
a created_at - what would that even mean?! They're not their own
objects, they're just in the deployment's plan. And so it was just
always a null. `is_cascading` however is more interesting.

This is so that the workflows from the REST response, are more similar
to the field returned by dsl-parser's prepare_plan
(it's still not the same, because dsl-parser returns a dict, and
this is a list, but changing that would break backcompat some)