RD-3031 Extract update steps in the new dep-update workflow (#3218)
* step-extractor: drop unused imports

huh, flake didnt scream about those? I think we dont flake dep-update/

* Copy step-extractor to the workflow

Not removing the original just yet. That'll come later.

* step-extractor: enable flake

yeah it was disabled

* step-extractor: drop disused functions

* REST: make the workflows part of the deployment response make sense

the workflow objects inside of deployments don't even contain
a created_at - what would that even mean?! They're not their own
objects, they're just in the deployment's plan. And so it was just
always a null. `is_cascading` however is more interesting.

This is so that the workflows from the REST response, are more similar
to the field returned by dsl-parser's prepare_plan
(it's still not the same, because dsl-parser returns a dict, and
this is a list, but changing that would break backcompat some)

* RD-3031 Call step-extractor from the workflow

A separate operation will extract the steps and store them in
the dep-update object.

* NO! BAD TEST!

Look, it's always None: https://github.com/cloudify-cosmo/cloudify-manager/blob/5ff42f1a717fc520ebb2f05c2236cde447212857/rest-service/manager_rest/storage/resource_models.py#L541
Stop having it.

* typo