RD-3085 + RD-3086 dep-up: apply changes to nodes+instances (#3233)
* RD-3085 + RD-3086 Store node/instance changes on the dep-up

Figure out the changes, and store them, to be applied "later".

Based on the steps, find nodes that were added, and nodes that
were modified. And put that into dep_up.nodes

* Update existing nodes

When performing the update, actually do update the changed nodes,
based on the attributes that have changed, which we know from
the code in the previous commit, and the new deployment plan.

* Also update node-instances

Update node-instances, mostly only wrt. their relationships. Similar
to the previous commit, but for instances, not nodes.

* Run the install/establish graphs

We've changed the nodes and instances, and so let's also actually
run the operations!