tenant query property
Note, this commit is going to be used by the ones after it, but TBH
it could well be a separate PR, because it's a Thing Unto Itself
as well.

This is something I wanted to do for a long time!
The overall idea is that I hate storage_manager being the only
interface for querying, and `sm._add_tenant_filter` means that the
only public way of having a tenant filter, is going through SM.
Instead, I add a public way of adding a tenant filter, for any query.
SM will delegate to that. And now any query can use it as well.