move all the context-rendering into exc.render_context
This is an important part, pay close attention! (hi, The Prodigy)

Move rendering the context into a method on the execution itself.
The idea is that exc.render_context() gives you a "bottled message"
that you render in a transaction, and pass to
`workflow_executor.execute_workflow` at any later time.

The important change is that the token is now set HERE. Before, we
only set it in execute_workflow itself, which meant that sending
executions was weird about transactions, because the token must
absolutely be committed before the executions are sent (otherwise,
the mgmtworker can try and run it before the token is actually stored,
and get a 401 immediately).

Now, you can prepare all executions up-front, bottling their messages
with tokens, commit, and only then actually send them. It should be
way easier to understand wrt. what happens at what point.

Oh, and object_session is going to be the same as just db.session in
most cases, but if we eventually started to use separate sessions
somewhere, this method would still continue to work then.