executions: split prepare/execute
This one's important and the very core of the idea.

Instead of having, on the resource_manager, .execute_workflow, only
have a .prepare_executions, which emits the execution messages,
after doing all the checks it deems necessary.

That can (and should) be done in the same transaction as all the prior
work and preparation.

So most of what this commits does, is replacing all rm.execute_workflow
calls with rm.prepare_executions+wf_exc.execute_workflow calls.
In the next commits, we will apply changes around those places, to
put transactions where proper, etc.

Then, you can pass those messages into
`workflow_executor.execute_workflow`, outside if a transaction.

It emits a list, because:
1. It accepts a list of executions :)
2. Any execution could be split into multiple messages, in the case of
   components.