separate transactions for dequeueing
This one is simple, but important!

The main part is mostly just switching the `while True` with the
`with transaction()` ordering:
`_get_queued_executions` will get a batch (of 5) executions to dequeue,
and if needed, then another batch, etc, until it's finally satisfied.

However, we learned that getting (and locking) separate batches in
a single transaction is a huge no-no, because if you have several
instances of that, then the other instance could also want several
batches, in the opposite order, leading to a deadlock.

Instead, get (and render) each batch in a separate transaction.
And send them at the very end.