dont check active executions if the deployment is not persisted
if the execution doesnt have a dep_fk, that means the deployment
is not persisted yet (and didnt get a PK). So we're in the dep-env
create execution. And no need to check for other executions of that
same deployment, because there'll be none of those anyway - the
deployment isn't created yet...