add ensure_defaults for executions
For when you want the defaults to be applied right here right now.
This is going to be necessary for things like execution.id, where we
want to render a message for an execution that is to-be committed,
so we do want the id (uuid) to be rendered right now.