kill-cancel: don't send an execution-token
Setting a new exec-token on the execution, when sending this message,
means that the original execution's token is effectively revoked,
because it's changed now.
No need to do it. The execution is going to be killed anyway.

If we do it, we get additional pointless 401s in the meantime.
It also means the original execution cannot possibly clean up
after itself (and it DOES get a chance to, remember it's a
TERM signal and then after a few seconds a KILL)