rewrite the check in _check_for_active_executions
changing the filter for a function is the simple part, it's just
handling the case where the execution is not created/stored yet,
and then, no need to have the additional SQL-side filter, since we
already know that all other executions will be older than it.

Rewrite the caller method a bit as well, wrt. making the `ifs` easier.
Also, if the deployment is not stored, no need to check for executions
of it: none can possibly have been run, if the deployment is not
stored yet.