commit after parts that don't always have an explicit transaction
resume and system workflows, we dont always call those from under
explicit transactions, so commit explicitly right here.

This will be able to be removed, when we do a similar split for
those two cases, as we did for prepare_execution+execute, where
perform (and then: execute-system, and resume) are under the
transaction, and the amqp message is not.
But not yet.