RD-3121 delete_blueprint: handle blueprints w/o imports
Old blueprints, like from a 4.5 snapshot, don't have this field
at all, and then the .get() returns None, and a None is not a good
target for a `x in None` check. Needs to default to something
that can be a container.

This fixes `integration_tests/tests/agentless_tests/test_snapshot.py::TestSnapshot::test_v_4_5_restore_snapshot_without_imported_blueprints`