keep the same error handling
Still - any exception thrown in those operations should just be
nonrecoverable - as it was before. But now that we don't throw
nonrecoverable in a deep point in the call stack, we need to just
catch errors at the top level, and re-throw them nonrecoverable.

This means all the code inside of the operations can now use
regular exception classes, not just nonrecoverable everything.