RD-3183 Refactor cloudify_types/ (#3264)
* Inline the component.upload_blueprint operation

* Inline the component.create operation

* Inline the component.delete operation

* Inline the component.execute_start operation

* Inline the shared_resource.connect_deployment operation

* Inline the shared_resource.disconnect_deployment operation

* operations: use the imported ctx

for consistency with other functions.

* Update calling convention in unittests

Also update mocks.

* drop this stupid test

all it does is download a file from https, accessing the internet
for no good reason, and then asserts that something is true, except
it's not true, it's a filename. Yes, we get it, requests.get works.

Sans this test, the whole suite runs almost immediately

* tests: s/assertRaisesRegexp/assertRaisesRegex/

drops some warnings

* s/logger.warn/logger.warning/

the old one, .warn, is deprecated.

* drop the notion of _http_client_wrapper

it's so pointless, man. Let's just not have it.

* dep-create: this throws cloudifyclienterror

now that we got rid of http_client_wrapper, we can work with actual
exception types, not NonRecoverable.

* keep the same error handling

Still - any exception thrown in those operations should just be
nonrecoverable - as it was before. But now that we don't throw
nonrecoverable in a deep point in the call stack, we need to just
catch errors at the top level, and re-throw them nonrecoverable.

This means all the code inside of the operations can now use
regular exception classes, not just nonrecoverable everything.

* Update tests with the new error handling

The messages are now, I'd say, better. And you can see the traceback.
Amazing.

* tests: update mocks to the new module

stop using the component/sharedresource classes in tests

The changes are trivial here, the biggest one is replacing the
Component().upload_plugins() call with a _upload_plugins() call

* drop Component/SharedResource classes

booom

* style changes for logs and exceptions

%s-logging for ctx.logger, f''-strings for exceptions

* Clear up the poll logging a bit

I don't want to see raw events/executions in the log. That's just
unreadable.

* component.create: factor out a function

move _do_create_deployment outside

* component tests: use regular resume, not force-resume

no reason to force-resume here. Force-resume is for when operations
were killed before

* do_create_deployment: pass async_create

we already wait for create-dep-env to finish (in the create operation)
so no need for sync-create here, let's async

* flake

* s/_verify_secrets_clash/_abort_if_secrets_clash/

* tests: handle no .status_display entry

...either in old client, or in tests with mocked executions as dicts