RD-3235 Plugins update: don't delete deployments on timeout (#3273)
* RD-3235 Plugins update: don't delete deployments on timeout

When the dep-upd inside of plug-upd times out, we'll continue on
to setting the plug-upd execution's state to FAILED, and we can't
just delete the temp blueprint! If the dep-upd started, deployments
might be using it!

First, set all deployments to use the original blueprint, instead.
If (by race) some deployments still use the temp one, then don't
delete it anyway, just un-hide it so that the user has a chance to
clean up.

* Also bump the plug-update timeout to 1 hour

Why not! The default 900 seconds seems to be just not enough