RD-3183 Make component.create_deployment resumable (#3278)
* Factor out _wait_for_deployment_create

* factor out _create_inter_deployment_dependency

* inline this call

* rewrite _do_create_deployment wrt. deployment id

I prefer it this way :)
Have a single approach to creating deployments, just switch on
the auto_inc_suffix inside of a function that creates the id.

* inline update_runtime_properties

this function is kinda useless, isn't it

* remove disused function

* _do_create_deployment: store created deployment id

when we do send the create-dep request sucessfully, store the
id. And, before sending, check if we already did it. That means
that resumes (ie. running again) won't try to create the deployment
again.

* also gate create_idd

Same way. Avoid running the function if we ran it before.

* factor the gating out to a utils function

* simplify _do_create_deployment

* _wait_for_deployment_create: use .create_execution

now this is an entry on the deployment itself, no need to search for
the execution

* deployments: allow _include=create_execution

* Don't break on None create-execution

* add the newly-returned field to the assert

* fixup component tests

now that we use a .get, it needs to be mocked out too.