RD-3264 step_extractor: disambiguate identical relationships (#3281)
* drop disused functions

I thought I've dropped these before, but alas nope

* _get_matching_relationship: only take type/target

don't take the whole relationship as an argument, because we only
filter on those two items, so taking the whole item, suggests that
we can also look at things like operations etc. Which we won't. So
don't suggest that.

* _get_matching_relationship: add a skip argument

For disambiguating multiple identical relationships: skip=0 means get
the first one, skip=1 means get the second matching one, etc.

* rename _get_matching_relationship to _find_relationship

this name fits better with me. Also add a docstring

* update unittests

..to use the new _find_relationship interface

* add a unittest

before the changes, it used to return a step that wanted to move
the rel from posiion 1 to position 0.

* also implement the same changes for the new dep-update extractor

this is still just a copy, before we remove the old one. So the
changes are copied too.

* modify-relationship inte-test: have two identical relationships

this happens in the wild, so let's test that we don't break then