RD-3094 new dep-update: deleting nodes (#3284)
* Factor out the non-preview function

Makes for less indentation in the big part of the update

* inline _install_new_instances after all

the install list will be amended by other things as well, so it's
going to be easier just having a single location at first

* Removal inte-tests

Similar to the other modules, and again starting with just a few
testcases. Not all of them are green just yet.

* RD-3094 Nodes DELETE endpoint

Allow deleting a node by its id

* RD-3094 Delete removed nodes

A whole part of the graph to do post-workflow stuff, in this case
deleting nodes. It's kinda similar to adding, but the other
way around!

* Uninstall deleted nodes

Also similar to installing, but again, the other way around!

* flake