Snapshots: omit audit-log entries
There's always going to be audit-log entries on the manager. This means
that restoring more from the snapshot, is going to always be restoring
to essentially a non-empty manager, and we're always going to run the
risk of storage id conflicts.

Unfortunate.

Well, let's not have audit-log entries in the snapshot, then.